feat: Allow user to override kysely plugins in SqliteDurableObject (#876)
[Context from Discord](https://discord.com/channels/679514959968993311/1307974274145062912/1436932741777064036)

I was having trouble creating a Better Auth adapter for DurableObject SQLite. It turned out that was because `SqliteDurableObject` from rwsdk automatically uses `ParseJSONResultsPlugin`. This PR allows the list of Kysely plugins to be set in the `SqliteDurableObject` constructor. 

I figured this was more scalable then just having a boolean flag to control the JSON parsing, but I'm happy to just use the boolean flag if that is preferable. 

cc: @peterp @justinvdm